Updated intc types

+++ src/hal/intc.h
@@ -19,7 +19,7 @@ typedef uint32 intc_id;
 
 void intc_initialize();
 
-void intc_add_handler(intc_id id, intc_handler handler, int aff_cpu);
+void intc_add_handler(intc_id id, intc_handler handler, uint8 aff_cpu);
 
 void intc_add_local_handler(intc_id id, intc_handler handler);
